Validation for Category/Products

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
+use App\Http\Requests\CategoryRequest;
 use App\Models\Category;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
@@ -29,11 +30,14 @@ class CategoryController extends Controller
     /**
      * Store a newly created resource in storage.
      */
-    public function store(Request $request)
+    public function store(CategoryRequest $request)
     {
-        $path = $request->file('image')->store('categories');
         $params = $request->all();
-        $params['image'] = $path;
+        unset($params['image']);
+        if ($request->has('image')) {
+            $filename = $request->file('image')->getClientOriginalName();
+            $params['image'] = $request->file('image')->storeAs('categories', $filename);
+        }
         Category::create($params);
         return redirect()->route('categories.index');
     }
@@ -57,12 +61,15 @@ class CategoryController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, Category $category)
+    public function update(CategoryRequest $request, Category $category)
     {
-        Storage::delete($category->image);
-        $path = $request->file('image')->store('categories');
         $params = $request->all();
-        $params['image'] = $path;
+        unset($params['image']);
+        if ($request->has('image')) {
+            Storage::delete($category->image);
+            $filename = $request->file('image')->getClientOriginalName();
+            $params['image'] = $request->file('image')->storeAs('categories', $filename);
+        }
         $category->update($params);
         return redirect()->route('categories.index');
     }
